Updates to the-essentials.sh
Added checks for several programs to check if they are updated, if they are updated, to check on github for versions and if they did not match, to pull the new binarys down.

Removed legacy ways of looking for binaries using "wget -qO-" for most of the items since github changed the way they are indexing items.
Example of this can be seen for the "install_aquatone" function.  

Will be adding more of these the further down the line.  I also noticed that i was not installing jq early enough for new users for the new methods i was using so i fixed those.
Fixed other installer issues i noticed from updates on other githus.  Overall cleanup and maintenance.


Cheers